+++ src/main/java/com/bookstore/Service/OrderItemService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface OrderItemService {
     List<OrderItem> findOrderItemsByOrderid(Long orderid);
 
-    void addNewItem(int orderid, String book, double price, int number);
+    void addNewItem(Long orderid, String book, double price, int number);
 }
